refactor: move selectors to selectors.js
