+++ src/features/categorySlice.js
@@ -135,14 +135,4 @@ export const {
   clearPoints,
 } = categorySlice.actions;
 
-// ? selectors
-export const selectTableData = (state) => state.tableData;
-export const selectSelectedClue = (state) => state.selectedClue;
-export const selectIsModalOpen = (state) => state.isModalOpen;
-export const selectClickedClues = (state) => state.clueClicked;
-export const selectPoints = (state) => state.points;
-export const selectIsAnswerCorrect = (state) => state.isAnswerCorrect;
-export const selectUser = (state) => state.user;
-export const selectStatus = (state) => state.status;
-
 export default categorySlice.reducer;
