+++ src/features/selectors.js
@@ -1,0 +1,9 @@
+// ? selectors
+export const selectTableData = (state) => state.tableData;
+export const selectSelectedClue = (state) => state.selectedClue;
+export const selectIsModalOpen = (state) => state.isModalOpen;
+export const selectClickedClues = (state) => state.clueClicked;
+export const selectPoints = (state) => state.points;
+export const selectIsAnswerCorrect = (state) => state.isAnswerCorrect;
+export const selectUser = (state) => state.user;
+export const selectStatus = (state) => state.status;
